Marklab/web/requirements.txt
Django~=5.2
psycopg2-binary~=2.9.10
python-dotenv~=1.1.0
docker~=7.1.0
paramiko~=3.5.1
djangorestframework~=3.16.0
gunicorn~=23.0.0
